@@ -4,16 +4,16 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.4",
+        "php": ">=8.5",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "guzzlehttp/guzzle": "^7.10",
-        "symfony/console": "7.3.*",
-        "symfony/dotenv": "7.3.*",
-        "symfony/flex": "^2.10",
-        "symfony/framework-bundle": "7.3.*",
-        "symfony/runtime": "7.3.*",
-        "symfony/yaml": "7.3.*"
+        "symfony/console": ">=8.0.3",
+        "symfony/dotenv": ">=8",
+        "symfony/flex": ">=2.10",
+        "symfony/framework-bundle": ">=8.0.3",
+        "symfony/runtime": ">=8.0.1",
+        "symfony/yaml": ">=8.0.1"
     },
     "config": {
         "allow-plugins": {
@@ -63,16 +63,17 @@
     "extra": {
         "symfony": {
             "allow-contrib": true,
-            "require": "7.3.*"
+            "require": "8.*"
         }
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^3.92",
+        "friendsofphp/php-cs-fixer": "^3.92.5",
         "phpstan/extension-installer": "^1.4.3",
         "phpstan/phpstan-symfony": "^2.0.9",
         "phpunit/phpunit": "^12.5.4",
-        "symfony/browser-kit": "7.3.*",
-        "symfony/css-selector": "7.3.*",
-        "symfony/phpunit-bridge": "^8.0"
+        "rector/rector": "^2.3",
+        "symfony/browser-kit": ">=8.0.3",
+        "symfony/css-selector": ">=8",
+        "symfony/phpunit-bridge": ">=8.0.3"
     }
 }
