@@ -4,6 +4,7 @@ use App\Kernel;
 
 require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
 
-return function (array $context) {
-    return new Kernel($context['APP_ENV'], (bool) $context['APP_DEBUG']);
-};
+return fn(array $context) => new Kernel(
+    is_string($context['APP_ENV']) ? $context['APP_ENV'] : 'prod',
+    (bool) $context['APP_DEBUG']
+);
